Fix(docat): Upload of file with name of tag
Uploading a document failed with the same name as a broken (no
reference) tag resulted in a 500 status code and a stacktrace. To fix
this, we now check if a symlink with the same name as the uploaded
version exists, and fail if it does.

We now also return a warning when a tag for a non existing version is created.

fixes #124